// src/models/enums/ResponsabilidadesIDs.tsx
enum ResponsabilidadesIDs {
  /* DiretorDEIN = 'dir-dein',
   * CoordenadorCOPR = 'coord-copr',
   * GerenteGEIN = 'ger-gein',
   * GerenteGETD = 'ger-getd',
   * CoordenadorCOPI = 'coord-copi',
   * CoordenadorCOGI = 'coord-cogi', */
  DEIN = 'dein',
  COPR = 'copr',
  GEST = 'gest',
  GEIN = 'gein',
  GENN = 'genn',
  GETD = 'getd',
  COPI = 'copi',
  COGI = 'cogi',
  COPD = 'copd',
  COTD = 'cotd',
  CODS = 'cods',
  CanaisParcerias = 'canais-parcerias',
  GestaoPortfolio = 'gestao-portfolio',
  ContratosAuditoria = 'contratos-auditorias',
  PagamentosCadastro = 'pagamentos-cadastro-nps',
  ComprasViagens = 'compras-viagens-folhas',
  GestaoOrcamentaria = 'gestao-orcamentaria',
  GestaoDesempenho = 'gestao-desempenho',
  GestaoGovernanca = 'gestao-governanca',
  RelacionamentoEcossistema = 'relacionamento-ecossistema',
  RelacionamentoInterno = 'relacionamento-interno',
  GestaoProjetos = 'gestao-projetos',
  OrcamentoGCP = 'orcamento-gpd',
  ENOVA = 'enova'
}

export default ResponsabilidadesIDs;